requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
gTTS>=2.5.0
groq>=0.4.0
Pillow>=10.0.0
django-cors-headers>=4.3.0
boto3>=1.29.0
dj-database-url>=2.1.0
mysqlclient>=2.2.0

